Refactoring hdemucs
