--- clarity/utils/__init__.py
@@ -1,3 +1,0 @@
-from clarity.utils.source_separation import separate_hdemucs
-
-__all__ = ["separate_hdemucs"]
